Separate the deployment of contrats and the tests
