+++ test/Proxy/UpgradeProxy.test.js
@@ -1,12 +1,12 @@
 const { expectEvent, expectRevert } = require('@openzeppelin/test-helpers')
 const { should } = require('chai').should()
 
-const { deployProxy, upgradeProxy  } = require('@openzeppelin/truffle-upgrades')
+const { deployProxy, upgradeProxy } = require('@openzeppelin/truffle-upgrades')
 const CMTAT1 = artifacts.require('CMTAT')
 const CMTAT2 = artifacts.require('CMTAT')
 
-contract("UpgradeableCMTAT - Proxy", function ([_, owner, address1]) {
-  it("should increment the balance value", async function () {
+contract('UpgradeableCMTAT - Proxy', function ([_, owner, address1]) {
+  it('should increment the balance value', async function () {
     // With the first version of CMTAT
     this.upgradeableCMTATInstance = await deployProxy(CMTAT1, [owner, 'CMTA Token', 'CMTAT', 'CMTAT_ISIN', 'https://cmta.ch'], { initializer: 'initialize', constructorArgs: [_] });
 
@@ -17,11 +17,11 @@ contract("UpgradeableCMTAT - Proxy", function ([_, owner, address1]) {
       from: owner
     }));
     (await this.upgradeableCMTATInstance.balanceOf(address1)).should.be.bignumber.equal('20');
-    (await this.upgradeableCMTATInstance.totalSupply()).should.be.bignumber.equal('20');
+    (await this.upgradeableCMTATInstance.totalSupply()).should.be.bignumber.equal('20')
 
     // With the new version
     // With the first version of CMTAT
-    //this.upgradeableCMTATV2Instance = await upgradeProxy(this.upgradeableCMTATInstance.address, CMTAT2, [owner, 'CMTA Token', 'CMTAT', 'CMTAT_ISIN', 'https://cmta.ch']);
+    // this.upgradeableCMTATV2Instance = await upgradeProxy(this.upgradeableCMTATInstance.address, CMTAT2, [owner, 'CMTA Token', 'CMTAT', 'CMTAT_ISIN', 'https://cmta.ch']);
     this.upgradeableCMTATV2Instance = await upgradeProxy(this.upgradeableCMTATInstance.address, CMTAT2, { constructorArgs: [_] });
 
     (await this.upgradeableCMTATV2Instance.balanceOf(address1)).should.be.bignumber.equal('20');
@@ -31,6 +31,6 @@ contract("UpgradeableCMTAT - Proxy", function ([_, owner, address1]) {
       from: owner
     }));
     (await this.upgradeableCMTATV2Instance.balanceOf(address1)).should.be.bignumber.equal('40');
-    (await this.upgradeableCMTATV2Instance.totalSupply()).should.be.bignumber.equal('40');
-    })
+    (await this.upgradeableCMTATV2Instance.totalSupply()).should.be.bignumber.equal('40')
+  })
 })
